feat(students statistics): show my students statistics by default
