+++ app/views/course/statistics/aggregate/all_students.json.jbuilder
@@ -5,26 +5,28 @@ course_video_count = has_course_videos ? course_videos.count : 0
 can_analyze_videos = can?(:analyze_videos, current_course)
 is_course_gamified = current_course.gamified?
 no_group_managers = @service.no_group_managers?
-
-json.metadata do
-  json.isCourseGamified is_course_gamified
-  json.showVideo has_course_videos && can_analyze_videos
-  json.courseVideoCount course_video_count
-  json.hasGroupManagers !no_group_managers
-end
+has_my_students = false
 
 json.students @all_students do |student|
+  is_my_student = false
   json.name student.name
   json.nameLink course_user_path(current_course, student)
   json.studentType student.phantom? ? 'Phantom' : 'Normal'
 
   unless no_group_managers
     json.groupManagers @service.group_managers_of(student) do |manager|
+      if manager.id == current_course_user&.id
+        is_my_student = true
+        has_my_students = true
+      end
+
       json.id manager.id
       json.name manager.name
       json.nameLink course_user_path(current_course, manager)
     end
   end
+
+  json.isMyStudent is_my_student
 
   if is_course_gamified
     json.level student.level_number
@@ -36,4 +38,12 @@ json.students @all_students do |student|
     json.videoSubmissionLink course_user_video_submissions_path(current_course, student)
     json.videoPercentWatched student.video_percent_watched
   end
+end
+
+json.metadata do
+  json.isCourseGamified is_course_gamified
+  json.showVideo has_course_videos && can_analyze_videos
+  json.courseVideoCount course_video_count
+  json.hasGroupManagers !no_group_managers
+  json.hasMyStudents has_my_students
 end
