+++ client/app/bundles/course/statistics/pages/StatisticsIndex/students/index.jsx
@@ -1,31 +1,12 @@
 import { defineMessages, useIntl } from 'react-intl';
-import Box from '@mui/material/Box';
-import LinearProgress from '@mui/material/LinearProgress';
-import Typography from '@mui/material/Typography';
-import PropTypes from 'prop-types';
 
 import ErrorCard from 'lib/components/core/ErrorCard';
-import DataTable from 'lib/components/core/layouts/DataTable';
-import Link from 'lib/components/core/Link';
 import LoadingIndicator from 'lib/components/core/LoadingIndicator';
-import { DEFAULT_TABLE_ROWS_PER_PAGE } from 'lib/constants/sharedConstants';
+import Note from 'lib/components/core/Note';
 
 import { studentsIndexShape } from '../../../propTypes/students';
 
-const options = {
-  downloadOptions: {
-    filename: 'students_statistics',
-  },
-  jumpToPage: true,
-  print: false,
-  rowsPerPage: DEFAULT_TABLE_ROWS_PER_PAGE,
-  rowsPerPageOptions: [DEFAULT_TABLE_ROWS_PER_PAGE],
-  selectableRows: 'none',
-  sortOrder: {
-    name: 'experiencePoints',
-    direction: 'desc',
-  },
-};
+import StudentsStatisticsTable from './StudentsStatisticsTable';
 
 const translations = defineMessages({
   error: {
@@ -33,67 +14,13 @@ const translations = defineMessages({
     defaultMessage:
       'Something went wrong when fetching student statistics! Please refresh to try again.',
   },
-  name: {
-    id: 'course.statistics.StatisticsIndex.students.name',
-    defaultMessage: 'Name',
-  },
-  studentType: {
-    id: 'course.statistics.StatisticsIndex.students.studentsType',
-    defaultMessage: 'Student Type',
-  },
-  groupManagers: {
-    id: 'course.statistics.StatisticsIndex.students.groupManagers',
-    defaultMessage: 'Tutors',
-  },
-  level: {
-    id: 'course.statistics.StatisticsIndex.students.level',
-    defaultMessage: 'Level',
-  },
-  experiencePoints: {
-    id: 'course.statistics.StatisticsIndex.students.experiencePoints',
-    defaultMessage: 'Experience Points',
-  },
-  videoSubmissionCount: {
-    id: 'course.statistics.StatisticsIndex.students.videoSubmissionCount',
-    defaultMessage: 'Videos Watched (Total: {courseVideoCount})',
-  },
-  videoPercentWatched: {
-    id: 'course.statistics.StatisticsIndex.students.videoPercentWatched',
-    defaultMessage: 'Average % Watched',
-  },
-  tableTitle: {
-    id: 'course.statistics.StatisticsIndex.students.tableTitle',
-    defaultMessage: 'Student Statistics',
-  },
-  tutorFilter: {
-    id: 'course.statistics.StatisticsIndex.students.tutorFilter',
-    defaultMessage: 'Tutor: {name}',
+  noStudents: {
+    id: 'course.statistics.StatisticsIndex.students.noStudents',
+    defaultMessage: 'There is no student in this course, yet...',
   },
 });
 
-const LinearProgressWithLabel = (props) => (
-  <Box sx={{ display: 'flex', alignItems: 'center' }}>
-    <Box sx={{ width: '100%', mr: 1 }}>
-      <LinearProgress variant="determinate" {...props} />
-    </Box>
-    <Box sx={{ minWidth: 35 }}>
-      <Typography color="text.secondary" variant="body2">{`${Math.round(
-        props.value ?? 0,
-      )}%`}</Typography>
-    </Box>
-  </Box>
-);
-
-LinearProgressWithLabel.propTypes = {
-  /**
-   * Value between 0 and 100.
-   */
-  value: PropTypes.number,
-};
-
 const StudentsStatistics = ({ metadata, students, isFetching, isError }) => {
-  const { isCourseGamified, showVideo, courseVideoCount, hasGroupManagers } =
-    metadata;
   const intl = useIntl();
 
   if (isFetching) {
@@ -102,154 +29,11 @@ const StudentsStatistics = ({ metadata, students, isFetching, isError }) => {
   if (isError) {
     return <ErrorCard message={intl.formatMessage(translations.error)} />;
   }
-  const columns = [
-    {
-      name: 'name',
-      label: intl.formatMessage(translations.name),
-      options: {
-        filter: false,
-        sort: true,
-      },
-    },
-    {
-      name: 'studentType',
-      label: intl.formatMessage(translations.studentType),
-      options: {
-        filter: false,
-        sort: true,
-      },
-    },
-  ];
-
-  if (hasGroupManagers) {
-    columns.push({
-      name: 'groupManagers',
-      label: intl.formatMessage(translations.groupManagers),
-      options: {
-        filter: true,
-        filterType: 'multiselect',
-        filterOptions: {
-          names: [
-            ...new Set(
-              students.flatMap((s) => s.groupManagers.map((m) => m.name)),
-            ),
-          ],
-          logic: (managers, filters) => {
-            if (filters) {
-              const filterSet = new Set(filters);
-              return !managers
-                .map((m) => m.name)
-                .some((name) => filterSet.has(name));
-            }
-            return false;
-          },
-          fullWidth: true,
-        },
-        customFilterListOptions: {
-          render: (name) =>
-            intl.formatMessage(translations.tutorFilter, { name }),
-        },
-        sort: true,
-        customBodyRenderLite: (dataIndex) => {
-          const groupManagers = students[dataIndex].groupManagers;
-          if (!groupManagers) {
-            return '';
-          }
-          return (
-            <>
-              {groupManagers.map((m, index) => (
-                <span key={m.id}>
-                  <Link href={m.nameLink} opensInNewTab>
-                    {m.name}
-                  </Link>
-                  {index < groupManagers.length - 1 && ', '}
-                </span>
-              ))}
-            </>
-          );
-        },
-      },
-    });
-  }
-  if (isCourseGamified) {
-    columns.push({
-      name: 'level',
-      label: intl.formatMessage(translations.level),
-      options: {
-        filter: true,
-        sort: true,
-      },
-    });
-    columns.push({
-      name: 'experiencePoints',
-      label: intl.formatMessage(translations.experiencePoints),
-      options: {
-        filter: false,
-        sort: true,
-        alignCenter: true,
-        customBodyRenderLite: (dataIndex) => {
-          const student = students[dataIndex];
-          return (
-            <Link
-              key={student.id}
-              href={student.experiencePointsLink}
-              opensInNewTab
-            >
-              {student.experiencePoints}
-            </Link>
-          );
-        },
-      },
-    });
-  }
-  if (showVideo) {
-    columns.push({
-      name: 'videoSubmissionCount',
-      label: intl.formatMessage(translations.videoSubmissionCount, {
-        courseVideoCount,
-      }),
-      options: {
-        alignCenter: true,
-        customBodyRenderLite: (dataIndex) => {
-          const student = students[dataIndex];
-          return (
-            <Link
-              key={student.id}
-              href={student.videoSubmissionLink}
-              opensInNewTab
-            >
-              {student.videoSubmissionCount}
-            </Link>
-          );
-        },
-      },
-    });
-    columns.push({
-      name: 'videoPercentWatched',
-      label: intl.formatMessage(translations.videoPercentWatched),
-      options: {
-        filter: false,
-        sort: true,
-        alignCenter: true,
-        customBodyRender: (value) => (
-          <Box sx={{ width: '100%' }}>
-            <LinearProgressWithLabel value={value} />
-          </Box>
-        ),
-      },
-    });
+  if (students.length === 0) {
+    return <Note message={intl.formatMessage(translations.error)} />;
   }
 
-  return (
-    <DataTable
-      columns={columns}
-      data={students}
-      height="30px"
-      includeRowNumber
-      options={options}
-      title={intl.formatMessage(translations.tableTitle)}
-    />
-  );
+  return <StudentsStatisticsTable metadata={metadata} students={students} />;
 };
 
 StudentsStatistics.propTypes = studentsIndexShape;
